requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
python-dotenv==1.0.1

